projects/product-search-api/requirements.txt
Flask==2.3.2
requests==2.31.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
